Refactor exceptions a bit
