+++ LibreNMS/Exceptions/ApiException.php
@@ -1,5 +1,5 @@
 <?php
-/**
+/*
  * ApiException.php
  *
  * -Description-
@@ -15,28 +15,40 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  *
- * @link       https://www.librenms.org
- *
- * @copyright  2019 Tony Murray
+ * @package    LibreNMS
+ * @link       http://librenms.org
+ * @copyright  2022 Tony Murray
  * @author     Tony Murray <murraytony@gmail.com>
  */
 
 namespace LibreNMS\Exceptions;
 
+use Illuminate\Http\JsonResponse;
+
 class ApiException extends \Exception
 {
-    private $output;
-
-    public function __construct($message = '', $output = [])
+    /**
+     * @param  string  $message
+     * @param  int  $code
+     * @param  \Throwable|null  $previous
+     */
+    public function __construct($message = '', $code = 400, $previous = null)
     {
-        parent::__construct($message, 0, null);
-        $this->output = $output;
+        parent::__construct($message, $code, $previous);
     }
 
-    public function getOutput()
+    /**
+     * Render the exception into an HTTP response.
+     *
+     * @param  \Illuminate\Http\Request  $request
+     */
+    public function render($request): JsonResponse
     {
-        return $this->output;
+        return response()->json([
+            'status'  => 'error',
+            'message' => $this->getMessage(),
+        ], $this->getCode(), [], JSON_PRETTY_PRINT);
     }
 }
